// src/app/shared/interfaces/Planet/planet.ts
export interface Planet {
  climate: string;
  diameter: string;
  filmIds: number[];
  films: string[];
  gravity: string;
  id: number;
  name: string;
  orbital_period: string;
  population: string;
  residentIds: number[];
  residents: string[];
  rotation_period: string;
  surface_water: string;
  terrain: string;
  url: string;
}